Refactored complexity report
