+++ detekt-cli/src/main/kotlin/io/gitlab/arturbosch/detekt/cli/console/ComplexityReport.kt
@@ -11,8 +11,8 @@ class ComplexityReport : ConsoleReport() {
 	override val priority: Int = 20
 
 	override fun render(detektion: Detektion): String? {
-      val reportGenerator = ReportGenerator.create(detektion)
-		  return reportGenerator.generate()
+		val complexityReportGenerator = ComplexityReportGenerator.create(detektion)
+		return complexityReportGenerator.generate()
 	}
 
 }
